added a struct to set the vehicle parameters and moved all the models to a specific folder as we are going to make more
