+++ code/system/application/sensation/include/vehicleModels/truckObservationModel.hpp
