+++ code/system/application/sensation/src/sensation.cpp
@@ -137,7 +137,6 @@ odcore::data::dmcp::ModuleExitCodeMessage::ModuleExitCode Sensation::body() {
     odcore::reflection::CSVFromVisitableVisitor csvExporter1(fout, WITH_HEADER, DELIMITER);
 
 
-
     double time_stamp = 0;
